+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -5,6 +5,9 @@ CHANGELOG
 -----
 
  * Added 'host' option to firewall configuration
+ * Added 'csrf_token_generator' and 'csrf_token_id' options to firewall logout
+   listener configuration to supercede/alias 'csrf_provider' and 'intention'
+   respectively
  * Moved 'security.secure_random' service configuration to FrameworkBundle
 
 2.3.0